fix: fix routing logic
