+++ src/routs/routs.ts
@@ -1,0 +1,6 @@
+enum ROUTS {
+  HOME = '/',
+  PRODUCT = '/product',
+}
+
+export { ROUTS };
